fix(filter-availability): only consider booked appointments when tracking clincian "commitments"
When computing the max availability we can display for a clinician, we do not
have to take into account how much _availability_ we've already shown for the
clinciain

If a patient is looking to book an appointment, they are looking to book a
single appointment at a time. Any of the available slots a clinician has are
fair game. It is only when a clinician reaches their limit for number of
appointments scheduled that we have to start excluding their available slots

Otherwise, we would only ever display the earliest set of appointmentments for
a clinician, never those in the afternoon (for example)